refactor(ai): move document loader to core/ai
