+++ packages/core/ai/src/index.ts
@@ -11,3 +11,4 @@ export * from './ai-manager';
 export * from './document-manager';
 export * from './tools-manager';
 export * from './loader';
+export * from './document-loader';
