+++ packages/plugins/@nocobase/plugin-ai/package.json
@@ -55,7 +55,6 @@
     "rehype-sanitize": "^6.0.0",
     "remark-gfm": "^4.0.1",
     "use-context-selector": "^2.0.0",
-    "xlsx": "^0.18.5",
     "zustand": "^5.0.6"
   },
   "keywords": [
